+++ src/sched.rs
@@ -100,6 +100,12 @@ mod sched_linux_like {
         }
     }
 
+    impl Default for CpuSet {
+        fn default() -> Self {
+            Self::new()
+        }
+    }
+
     /// `sched_setaffinity` set a thread's CPU affinity mask
     /// ([`sched_setaffinity(2)`](http://man7.org/linux/man-pages/man2/sched_setaffinity.2.html))
     ///