+++ CHANGELOG.md
@@ -35,6 +35,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   (#[1252](https://github.com/nix-rust/nix/pull/1252))
 - Added support for `Ipv4PacketInfo` and `Ipv6PacketInfo` to `ControlMessage`.
   (#[1222](https://github.com/nix-rust/nix/pull/1222))
+- `CpuSet` now implements `Default`.
+  (#[1244](https://github.com/nix-rust/nix/pull/1244))
 
 ### Changed
 - Changed `fallocate` return type from `c_int` to `()` (#[1201](https://github.com/nix-rust/nix/pull/1201))